Move parser state backup/restore into IParser API (#218)
* initial commit

* rename restoreCheckpoint -> restoreFromCheckpoint

* removed dead code
