+++ src/parser/IParser/IParserUtils.ts
@@ -6,8 +6,10 @@ import { CommonError, ResultUtils } from "../../common";
 import { Ast } from "../../language";
 import { LexerSnapshot } from "../../lexer";
 import { ParseSettings } from "../../settings";
+import { ParseContext, ParseContextUtils } from "../context";
 import { IParserState, IParserStateUtils } from "../IParserState";
-import { IParser, TriedParse } from "./IParser";
+import { NodeIdMap, NodeIdMapUtils } from "../nodeIdMap";
+import { IParser, IParserStateCheckpoint, TriedParse } from "./IParser";
 
 export function tryParse<S extends IParserState = IParserState>(
     parseSettings: ParseSettings<S>,
@@ -91,6 +93,64 @@ export function tryParseDocument<S extends IParserState = IParserState>(
 
             return ResultUtils.errFactory(ensureParseError(betterParsedState, betterParsedError, parseSettings.locale));
         }
+    }
+}
+
+// Due to performance reasons the backup no longer can include a naive deep copy of the context state.
+// Instead it's assumed that a backup is made immediately before a try/catch read block.
+// This means the state begins in a parsing context and the backup will either be immediately consumed or dropped.
+// Therefore we only care about the delta between before and after the try/catch block.
+// Thanks to the invariants above and the fact the ids for nodes are an auto-incrementing integer
+// we can easily just drop all delete all context nodes past the id of when the backup was created.
+export function stateCheckpointFactory(state: IParserState): IParserStateCheckpoint {
+    return {
+        tokenIndex: state.tokenIndex,
+        contextStateIdCounter: state.contextState.idCounter,
+        maybeContextNodeId: state.maybeCurrentContextNode?.id,
+    };
+}
+
+// See stateCheckpointFactory above for more information.
+export function restoreStateCheckpoint(state: IParserState, checkpoint: IParserStateCheckpoint): void {
+    state.tokenIndex = checkpoint.tokenIndex;
+    state.maybeCurrentToken = state.lexerSnapshot.tokens[state.tokenIndex];
+    state.maybeCurrentTokenKind = state.maybeCurrentToken?.kind;
+
+    const contextState: ParseContext.State = state.contextState;
+    const nodeIdMapCollection: NodeIdMap.Collection = state.contextState.nodeIdMapCollection;
+    const backupIdCounter: number = checkpoint.contextStateIdCounter;
+    contextState.idCounter = backupIdCounter;
+
+    const newContextNodeIds: number[] = [];
+    const newAstNodeIds: number[] = [];
+    for (const nodeId of nodeIdMapCollection.astNodeById.keys()) {
+        if (nodeId > backupIdCounter) {
+            newAstNodeIds.push(nodeId);
+        }
+    }
+    for (const nodeId of nodeIdMapCollection.contextNodeById.keys()) {
+        if (nodeId > backupIdCounter) {
+            newContextNodeIds.push(nodeId);
+        }
+    }
+
+    const sortByNumber: (left: number, right: number) => number = (left: number, right: number) => left - right;
+    for (const nodeId of newAstNodeIds.sort(sortByNumber).reverse()) {
+        const maybeParentId: number | undefined = nodeIdMapCollection.parentIdById.get(nodeId);
+        const parentWillBeDeleted: boolean = maybeParentId !== undefined && maybeParentId >= backupIdCounter;
+        ParseContextUtils.deleteAst(state.contextState, nodeId, parentWillBeDeleted);
+    }
+    for (const nodeId of newContextNodeIds.sort(sortByNumber).reverse()) {
+        ParseContextUtils.deleteContext(state.contextState, nodeId);
+    }
+
+    if (checkpoint.maybeContextNodeId) {
+        state.maybeCurrentContextNode = NodeIdMapUtils.assertGetContext(
+            state.contextState.nodeIdMapCollection.contextNodeById,
+            checkpoint.maybeContextNodeId,
+        );
+    } else {
+        state.maybeCurrentContextNode = undefined;
     }
 }
