+++ src/parser/IParserState/IParserStateUtils.ts
@@ -1,23 +1,12 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { NodeIdMap, ParseContext, ParseContextUtils, ParseError } from "..";
+import { ParseContext, ParseContextUtils, ParseError } from "..";
 import { Assert, CommonError, ICancellationToken } from "../../common";
 import { Ast, Constant, Token } from "../../language";
 import { LexerSnapshot } from "../../lexer";
 import { SequenceKind } from "../error";
-import { NodeIdMapUtils } from "../nodeIdMap";
 import { IParserState } from "./IParserState";
-
-export interface FastStateBackup {
-    readonly tokenIndex: number;
-    readonly contextStateIdCounter: number;
-    readonly maybeContextNodeId: number | undefined;
-}
-
-// ---------------------------
-// ---------- State ----------
-// ---------------------------
 
 // If you have a custom parser + parser state, then you'll have to create your own factory function.
 // See `benchmark.ts` for an example.
@@ -39,73 +28,6 @@ export function stateFactory(
         contextState: ParseContextUtils.stateFactory(),
         maybeCurrentContextNode: undefined,
     };
-}
-
-export function applyState(originalState: IParserState, otherState: IParserState): void {
-    originalState.tokenIndex = otherState.tokenIndex;
-    originalState.maybeCurrentToken = otherState.maybeCurrentToken;
-    originalState.maybeCurrentTokenKind = otherState.maybeCurrentTokenKind;
-
-    originalState.contextState = otherState.contextState;
-    originalState.maybeCurrentContextNode = otherState.maybeCurrentContextNode;
-}
-
-// Due to performance reasons the backup no longer can include a naive deep copy of the context state.
-// Instead it's assumed that a backup is made immediately before a try/catch read block.
-// This means the state begins in a parsing context and the backup will either be immediately consumed or dropped.
-// Therefore we only care about the delta between before and after the try/catch block.
-// Thanks to the invariants above and the fact the ids for nodes are an auto-incrementing integer
-// we can easily just drop all delete all context nodes past the id of when the backup was created.
-export function fastStateBackup(state: IParserState): FastStateBackup {
-    return {
-        tokenIndex: state.tokenIndex,
-        contextStateIdCounter: state.contextState.idCounter,
-        maybeContextNodeId: state.maybeCurrentContextNode?.id,
-    };
-}
-
-// See state.fastSnapshot for more information.
-export function applyFastStateBackup(state: IParserState, backup: FastStateBackup): void {
-    state.tokenIndex = backup.tokenIndex;
-    state.maybeCurrentToken = state.lexerSnapshot.tokens[state.tokenIndex];
-    state.maybeCurrentTokenKind = state.maybeCurrentToken?.kind;
-
-    const contextState: ParseContext.State = state.contextState;
-    const nodeIdMapCollection: NodeIdMap.Collection = state.contextState.nodeIdMapCollection;
-    const backupIdCounter: number = backup.contextStateIdCounter;
-    contextState.idCounter = backupIdCounter;
-
-    const newContextNodeIds: number[] = [];
-    const newAstNodeIds: number[] = [];
-    for (const nodeId of nodeIdMapCollection.astNodeById.keys()) {
-        if (nodeId > backupIdCounter) {
-            newAstNodeIds.push(nodeId);
-        }
-    }
-    for (const nodeId of nodeIdMapCollection.contextNodeById.keys()) {
-        if (nodeId > backupIdCounter) {
-            newContextNodeIds.push(nodeId);
-        }
-    }
-
-    const sortByNumber: (left: number, right: number) => number = (left: number, right: number) => left - right;
-    for (const nodeId of newAstNodeIds.sort(sortByNumber).reverse()) {
-        const maybeParentId: number | undefined = nodeIdMapCollection.parentIdById.get(nodeId);
-        const parentWillBeDeleted: boolean = maybeParentId !== undefined && maybeParentId >= backupIdCounter;
-        ParseContextUtils.deleteAst(state.contextState, nodeId, parentWillBeDeleted);
-    }
-    for (const nodeId of newContextNodeIds.sort(sortByNumber).reverse()) {
-        ParseContextUtils.deleteContext(state.contextState, nodeId);
-    }
-
-    if (backup.maybeContextNodeId) {
-        state.maybeCurrentContextNode = NodeIdMapUtils.assertGetContext(
-            state.contextState.nodeIdMapCollection.contextNodeById,
-            backup.maybeContextNodeId,
-        );
-    } else {
-        state.maybeCurrentContextNode = undefined;
-    }
 }
 
 export function startContext(state: IParserState, nodeKind: Ast.NodeKind): void {
